refactor: deduplicate code
