+++ src/index.ts
@@ -12,4 +12,4 @@ export { CONSTANTS } from './constants'
 export * from './flow'
 export * from './steps'
 export { default as VerificationStatus } from './verification-status'
-export { TinfoilBadge, type TinfoilBadgeElement } from './tinfoil-badge'
+export type { TinfoilBadgeElement } from './wc/tinfoil-badge'
